python: reorganize channel types
Move the remaining channel types out of channeltypes.py and into the
channels/ subdirectory.

Combine the fs* channels into a single file.

Drop the weird side-effecting decorator approach to building up the
CHANNEL_TYPES array.  We might bring it back later, but for now, use
some class variables on the channel subclass.
