+++ src/cockpit/channels/stream.py
@@ -1,3 +1,20 @@
+# This file is part of Cockpit.
+#
+# Copyright (C) 2022 Red Hat, Inc.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 import asyncio
 import logging
 import os
@@ -8,6 +25,8 @@ logger = logging.getLogger(__name__)
 
 
 class StreamChannel(AsyncChannel):
+    payload = 'stream'
+
     async def send_reader(self, reader):
         while True:
             logger.debug('waiting to read from process')
